update add fungible token
